Move getTableContentHeight out of the content model
