+++ src/components/tools/table-tool/table-tool.tsx
@@ -2,9 +2,10 @@ import { observer } from "mobx-react";
 import { onSnapshot } from "mobx-state-tree";
 import React, { useCallback, useEffect, useRef, useState } from "react";
 import ReactDataGrid from "react-data-grid";
-import { getTableContentHeight, TableContentModelType } from "../../../models/tools/table/table-content";
+import { TableContentModelType } from "../../../models/tools/table/table-content";
 import { exportTableContentAsJson } from "../../../models/tools/table/table-export";
 import { IToolTileProps } from "../tool-tile";
+import { getTableContentHeight } from "./table-utils";
 import { EditableTableTitle } from "./editable-table-title";
 import { TableToolbar } from "./table-toolbar";
 import { useColumnsFromDataSet } from "./use-columns-from-data-set";
